backend/requirements.txt
Flask==3.0.3
mysql-connector-python==8.0.33
python-dotenv==1.0.1
Flask-Mail==0.9.1
Flask-WTF==1.2.1
